src/backend/requirements.txt
# Flask, MySQL Connector
flask
mysql-connector-python
flask-cors
